refactor: move recording UI components to website example
Remove RecordButton, RecordingIndicator, MicrophoneSelector, and VUMeter
from @waveform-playlist/recording package. These are example-specific UI
components, not library concerns. RecordButton, RecordingIndicator, and
MicrophoneSelector now live in website/src/components/examples/RecordingControls.tsx.
VUMeter is replaced by SegmentedVUMeter from @waveform-playlist/ui-components.

Update docs to reference SegmentedVUMeter instead of the removed VUMeter.
