--- packages/recording/src/components/index.ts
@@ -1,15 +1,0 @@
-/**
- * Recording components exports
- */
-
-export { RecordButton } from './RecordButton';
-export type { RecordButtonProps } from './RecordButton';
-
-export { MicrophoneSelector } from './MicrophoneSelector';
-export type { MicrophoneSelectorProps } from './MicrophoneSelector';
-
-export { RecordingIndicator } from './RecordingIndicator';
-export type { RecordingIndicatorProps } from './RecordingIndicator';
-
-export { VUMeter } from './VUMeter';
-export type { VUMeterProps } from './VUMeter';
